Refactor inputs management
Creates tree new classes:
- InputManager
- Input
- InputAction

**InputManager**
You give it a list a key/action pairs and it'll create an input for each
and link them to the action.
Creates and manages the event listener.

**Input**
Each input tie to a different key and to an action.
It keeps track of when it's pressed.

**InputAction**
The final output (as multiple inputs can be linked to the same action,
it's important to listen to the `InputAction` and not on the `Input`).
Can have a callback which is a method who will be called when the action
is pressed.