[IMP] Add key mapping in settings
The key binding can be configured in the settings' menu.

If the user changes a key, it will be saved in the `localStorage`.
That way, next time the user wants to play, their key preference will be
found so they don't have to reconfigure the keys mapping.

Also, for now, there is only on key by action.
In the future, each action could have two key assigned to them.